+++ src/main/java/net/bios/frenchvanilla/NbtStrings.java
@@ -1,5 +1,0 @@
-package net.bios.frenchvanilla;
-
-public class NbtStrings {
-    public static final String MASS_STORAGE_CONVERTER = "frenchvanilla:mass_barrel_converter";
-}